Begin flattening
