--- src/st/prize_drop_fall.h
@@ -1,18 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-static void PrizeDropFall(void) {
-    if (g_CurrentEntity->velocityY >= 0) {
-        g_CurrentEntity->ext.equipItemDrop.fallSpeed +=
-            g_CurrentEntity->ext.equipItemDrop.gravity;
-        g_CurrentEntity->velocityX =
-            g_CurrentEntity->ext.equipItemDrop.fallSpeed;
-        if (g_CurrentEntity->velocityX == FIX(1) ||
-            g_CurrentEntity->velocityX == FIX(-1)) {
-            g_CurrentEntity->ext.equipItemDrop.gravity =
-                -g_CurrentEntity->ext.equipItemDrop.gravity;
-        }
-    }
-
-    if (g_CurrentEntity->velocityY < FIX(0.25)) {
-        g_CurrentEntity->velocityY += FIX(0.125);
-    }
-}
